Give the name to this fork and all the crates
We are going to use this for generating IDL in Light Protocol, so we
will most likely also release it on crates.io (and maybe npmjs.org) and
make Docker images based on it.

Therefore, to avoid name collision, we need to change the names of
crates in the least intrusive way. This change gives the `light-` prefix
to all crates which are going to be published. Using them with the
`package` option in Cargo.toml, like:

```
anchor-syn = { version = "0.27.0", package = "light-anchor-syn" }
```

is going to work as drop-in replacement for upstream Anchor without any
necessity to change import paths (`use anchor_syn`).

Also, mention the purpose of the fork in README.