IDL: Use cargo-expand to expand macros in code
cargo-expand is already used for the `anchor expand` subcommand.
In contrast to using `cargo +nightly rustc --profile=check --
-Zunpretty=expanded`, it works without problems in all of our
Anchor programs.

The code for using cargo-expand used to be a part of anchor-cli
crate, but since we had to use it in anchor-syn, this change moves
it to a separate crate.