IDL: Disable rustfmt when expanding the code to parse IDL from (#24)
Apparently, the newest rustfmt doesn't like the account constraint
syntax in Anchor.

This code:

```rust
\#[account(constraint = foo == bar @ ErrorCode::Baz)]
```

After rustfmt, becomes:

```rust
\#[account(constraint = foo = = bar @ ErrorCode::Baz)]
```

So it splits `==` to `= =`. syn is able to parse the former, but
not the latter.

This looks like a bug in rustfmt, but for now, let's disable it
as a workaround.